Move fixtures specific to integration/ to integration/.
