+++ robot-server/tests/conftest.py
@@ -1,22 +1,18 @@
-import signal
-import subprocess
-import time
-import sys
-import tempfile
-import os
 import json
+import os
 import pathlib
-import requests
-import pytest
-
+import tempfile
 from datetime import datetime, timezone
-from fastapi import routing
 from mock import MagicMock
-from starlette.testclient import TestClient
-from typing import Any, Callable, Generator, Iterator, cast
-from typing_extensions import NoReturn
 from pathlib import Path
+from typing import Callable, Generator, Iterator, cast
+from typing_extensions import NoReturn
+
 from sqlalchemy.engine import Engine as SQLEngine
+
+import pytest
+from fastapi import routing
+from starlette.testclient import TestClient
 
 from opentrons_shared_data.labware.dev_types import LabwareDefinition
 
@@ -134,72 +130,11 @@ def api_client_no_errors(
 
 
 @pytest.fixture(scope="session")
-def request_session() -> requests.Session:
-    session = requests.Session()
-    session.headers.update({API_VERSION_HEADER: LATEST_API_VERSION_HEADER_VALUE})
-    return session
-
-
-@pytest.fixture(scope="session")
 def server_temp_directory() -> Iterator[str]:
     new_dir = tempfile.mkdtemp()
     os.environ["OT_API_CONFIG_DIR"] = new_dir
     config.reload()
     yield new_dir
-
-
-@pytest.fixture(scope="session")
-def run_server(
-    request_session: requests.Session, server_temp_directory: str
-) -> Iterator["subprocess.Popen[Any]"]:
-    """Run the robot server in a background process."""
-    # In order to collect coverage we run using `coverage`.
-    # `-a` is to append to existing `.coverage` file.
-    # `--source` is the source code folder to collect coverage stats on.
-    with subprocess.Popen(
-        [
-            sys.executable,
-            "-m",
-            "coverage",
-            "run",
-            "-a",
-            "--source",
-            "robot_server",
-            "-m",
-            "uvicorn",
-            "robot_server:app",
-            "--host",
-            "localhost",
-            "--port",
-            "31950",
-        ],
-        env={
-            "OT_ROBOT_SERVER_DOT_ENV_PATH": "dev.env",
-            "OT_API_CONFIG_DIR": server_temp_directory,
-        },
-        stdin=subprocess.DEVNULL,
-        # The server will log to its stdout or stderr.
-        # Let it inherit our stdout and stderr so pytest captures its logs.
-        stdout=None,
-        stderr=None,
-    ) as proc:
-        # Wait for a bit to get started by polling /hcpealth
-        from requests.exceptions import ConnectionError
-
-        while True:
-            try:
-                request_session.get("http://localhost:31950/health")
-            except ConnectionError:
-                pass
-            else:
-                break
-            time.sleep(0.5)
-        request_session.post(
-            "http://localhost:31950/robot/home", json={"target": "robot"}
-        )
-        yield proc
-        proc.send_signal(signal.SIGTERM)
-        proc.wait()
 
 
 @pytest.fixture
@@ -259,20 +194,6 @@ def session_manager(hardware: HardwareControlAPI) -> SessionManager:
         hardware=hardware,
         motion_lock=ThreadedAsyncLock(),
     )
-
-
-@pytest.fixture
-def set_disable_fast_analysis(
-    request_session: requests.Session,
-) -> Iterator[None]:
-    """For integration tests that need to set then clear the
-    enableHttpProtocolSessions feature flag"""
-    url = "http://localhost:31950/settings"
-    data = {"id": "disableFastProtocolUpload", "value": True}
-    request_session.post(url, json=data)
-    yield None
-    data["value"] = None
-    request_session.post(url, json=data)
 
 
 @pytest.fixture
